+++ directdm/num/single_nucleon_form_factors.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python3
-""" The single-nucleon form factors """
 
-from directdm.num.num_input import Num_input 
+from directdm.num.num_input import Num_input
 
-class F1(object):
+class F1:
     def __init__(self, quark, nucleon):
         """ The nuclear form factor F1
-
+        
         Return the nuclear form factor F1
 
         Arguments
@@ -15,12 +14,12 @@ class F1(object):
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
         """
+
         self.quark = quark
         self.nucleon = nucleon
 
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
 
         if self.nucleon == 'p':
             if self.quark == 'u':
@@ -39,7 +38,7 @@ class F1(object):
 
 
 class F2(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor F2
 
         Return the nuclear form factor F2
@@ -49,32 +48,43 @@ class F2(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
 
+        ip = self.input_dict
+        
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return 2*ip.ap + ip.an + ip.F2sp
+                return 2*ip['ap'] + ip['an'] + ip['F2sp']
             if self.quark == 'd':
-                return 2*ip.an + ip.ap + ip.F2sp
+                return 2*ip['an'] + ip['ap'] + ip['F2sp']
             if self.quark == 's':
-                return ip.F2sp
+                return ip['F2sp']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return 2*ip.an + ip.ap + ip.F2sp
+                return 2*ip['an'] + ip['ap'] + ip['F2sp']
             if self.quark == 'd':
-                return 2*ip.ap + ip.an + ip.F2sp
+                return 2*ip['ap'] + ip['an'] + ip['F2sp']
             if self.quark == 's':
-                return ip.F2sp
+                return ip['F2sp']
 
 
 class FA(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FA at zero momentum transfer
 
         Return the nuclear form factor FA, evaluated at zero momentum transfer.
@@ -84,32 +94,42 @@ class FA(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
+        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.Deltaup
+                return ip['Deltaup']
             if self.quark == 'd':
-                return ip.Deltadp
+                return ip['Deltadp']
             if self.quark == 's':
-                return ip.Deltas
+                return ip['Deltas']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip.Deltaun
+                return ip['Deltaun']
             if self.quark == 'd':
-                return ip.Deltadn
+                return ip['Deltadn']
             if self.quark == 's':
-                return ip.Deltas
+                return ip['Deltas']
 
 
 class FPprimed(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FPprimed
 
         Return the nuclear form factor FPprimed
@@ -119,29 +139,40 @@ class FPprimed(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
+
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by 1 / (q^2 + mpi0^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.mN**2 * 2 * ip.gA
+                return self.mN**2 * 2 * ip['gA']
             if self.quark == 'd':
-                return - ip.mN**2 * 2 * ip.gA
+                return - self.mN**2 * 2 * ip['gA']
             if self.quark == 's':
                 return 0
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - ip.mN**2 * 2 * ip.gA
+                return - self.mN**2 * 2 * ip['gA']
             if self.quark == 'd':
-                return ip.mN**2 * 2 * ip.gA
+                return self.mN**2 * 2 * ip['gA']
             if self.quark == 's':
                 return 0
 
@@ -150,26 +181,27 @@ class FPprimed(object):
 
         The eta pole is given, in terms of the spatial momentum q, by 1 / (q^2 + meta^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.mN**2 * 2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
             if self.quark == 'd':
-                return ip.mN**2 * 2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
             if self.quark == 's':
-                return - ip.mN**2 * 4 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return - self.mN**2 * 4 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip.mN**2 * 2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
             if self.quark == 'd':
-                return ip.mN**2 * 2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return self.mN**2 * 2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
             if self.quark == 's':
-                return - ip.mN**2 * 4 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3
+                return - self.mN**2 * 4 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3
 
 
 class FS(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FS
 
         Return the nuclear form factor FS
@@ -179,32 +211,42 @@ class FS(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
+        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.sigmaup
+                return ip['sigmaup']
             if self.quark == 'd':
-                return ip.sigmadp
+                return ip['sigmadp']
             if self.quark == 's':
-                return ip.sigmas
+                return ip['sigmas']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip.sigmaun
+                return ip['sigmaun']
             if self.quark == 'd':
-                return ip.sigmadn
+                return ip['sigmadn']
             if self.quark == 's':
-                return ip.sigmas
+                return ip['sigmas']
 
 
 class FP(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FP
 
         Return the nuclear form factor FP
@@ -214,29 +256,40 @@ class FP(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
+
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by 1 / (q^2 + mpi0^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.mN**2 * ip.gA * ip.B0mu / ip.mN 
+                return self.mN**2 * ip['gA'] * ip['B0mu'] / self.mN 
             if self.quark == 'd':
-                return - ip.mN**2 * ip.gA * ip.B0md / ip.mN 
+                return - self.mN**2 * ip['gA'] * ip['B0md'] / self.mN 
             if self.quark == 's':
                 return 0
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - ip.mN**2 * ip.gA * ip.B0mu / ip.mN 
+                return - self.mN**2 * ip['gA'] * ip['B0mu'] / self.mN 
             if self.quark == 'd':
-                return ip.mN**2 * ip.gA * ip.B0md / ip.mN 
+                return self.mN**2 * ip['gA'] * ip['B0md'] / self.mN 
             if self.quark == 's':
                 return 0
 
@@ -245,26 +298,27 @@ class FP(object):
 
         The eta pole is given, in terms of the spatial momentum q, by 1 / (q^2 + meta^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0mu
+                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0mu']
             if self.quark == 'd':
-                return ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0md
+                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0md']
             if self.quark == 's':
-                return - 2 * ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0ms
+                return - 2 * self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0ms']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0mu
+                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0mu']
             if self.quark == 'd':
-                return ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0md
+                return self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0md']
             if self.quark == 's':
-                return - 2 * ip.mN**2 * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas)/3/ip.mN * ip.B0ms
+                return - 2 * self.mN**2 * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])/3/self.mN * ip['B0ms']
 
 
 class FT0(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FT0
 
         Return the nuclear form factor FT0
@@ -274,32 +328,42 @@ class FT0(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
+        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return ip.mu_at_2GeV * ip.gTu
+                return ip['mu_at_2GeV'] * ip['gTu']
             if self.quark == 'd':
-                return ip.md_at_2GeV * ip.gTd
+                return ip['md_at_2GeV'] * ip['gTd']
             if self.quark == 's':
-                return ip.ms_at_2GeV * ip.gTs
+                return ip['ms_at_2GeV'] * ip['gTs']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return ip.mu_at_2GeV * ip.gTd
+                return ip['mu_at_2GeV'] * ip['gTd']
             if self.quark == 'd':
-                return ip.md_at_2GeV * ip.gTu
+                return ip['md_at_2GeV'] * ip['gTu']
             if self.quark == 's':
-                return ip.ms_at_2GeV * ip.gTs
+                return ip['ms_at_2GeV'] * ip['gTs']
 
 
 class FT1(object):
-    def __init__(self, quark, nucleon):
+    def __init__(self, quark, nucleon, input_dict=None):
         """ The nuclear form factor FT1
 
         Return the nuclear form factor FT1
@@ -309,32 +373,42 @@ class FT1(object):
         quark = 'u', 'd', 's' -- the quark flavor (up, down, strange)
 
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.quark = quark
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
+        ip = self.input_dict
 
         if self.nucleon == 'p':
             if self.quark == 'u':
-                return - ip.mu_at_2GeV * ip.BT10up
+                return - ip['mu_at_2GeV'] * ip['BT10up']
             if self.quark == 'd':
-                return - ip.md_at_2GeV * ip.BT10dp
+                return - ip['md_at_2GeV'] * ip['BT10dp']
             if self.quark == 's':
-                return - ip.ms_at_2GeV * ip.BT10s
+                return - ip['ms_at_2GeV'] * ip['BT10s']
         if self.nucleon == 'n':
             if self.quark == 'u':
-                return - ip.mu_at_2GeV * ip.BT10un
+                return - ip['mu_at_2GeV'] * ip['BT10un']
             if self.quark == 'd':
-                return - ip.md_at_2GeV * ip.BT10dn
+                return - ip['md_at_2GeV'] * ip['BT10dn']
             if self.quark == 's':
-                return - ip.ms_at_2GeV * ip.BT10s
+                return - ip['ms_at_2GeV'] * ip['BT10s']
 
 
 class FG(object):
-    def __init__(self, nucleon):
+    def __init__(self, nucleon, input_dict=None):
         """ The nuclear form factor FG
 
         Return the nuclear form factor FG
@@ -342,21 +416,31 @@ class FG(object):
         Arguments
         ---------
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
+        ip = self.input_dict
 
         if self.nucleon == 'p':
-            return -2*ip.mG/27
+            return -2*ip['mG']/27
         if self.nucleon == 'n':
-            return -2*ip.mG/27
+            return -2*ip['mG']/27
 
 
 class FGtilde(object):
-    def __init__(self, nucleon):
+    def __init__(self, nucleon, input_dict=None):
         """ The nuclear form factor FGtilde
 
         Return the nuclear form factor FGtilde
@@ -364,40 +448,62 @@ class FGtilde(object):
         Arguments
         ---------
         nucleon = 'p', 'n' -- the nucleon (proton or neutron)
+
+        input_dict (optional) -- a dictionary of hadronic input parameters
+                                 (default is Num_input().input_parameters)
         """
         self.nucleon = nucleon
 
+        if input_dict is None:
+            self.input_dict = Num_input().input_parameters
+            # One should include a warning in case the dictionary
+            # does not contain all necessary keys
+        else:
+            self.input_dict = input_dict
+
     def value_zero_mom(self):
         """ Return the value of the form factor at zero momentum transfer """
-        ip = Num_input()
-
+        ip = self.input_dict
+        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
+        
         if self.nucleon == 'p':
-            return -ip.mN * ip.mtilde * (ip.Deltaup/ip.mu_at_2GeV + ip.Deltadp/ip.md_at_2GeV + ip.Deltas/ip.ms_at_2GeV)
+            return -self.mN * self.mtilde * (ip['Deltaup']/ip['mu_at_2GeV']\
+                                              + ip['Deltadp']/ip['md_at_2GeV']\
+                                              + ip['Deltas']/ip['ms_at_2GeV'])
         if self.nucleon == 'n':
-            return -ip.mN * ip.mtilde * (ip.Deltaup/ip.mu_at_2GeV + ip.Deltadp/ip.md_at_2GeV + ip.Deltas/ip.ms_at_2GeV)
+            return -self.mN * self.mtilde * (ip['Deltaup']/ip['mu_at_2GeV']\
+                                              + ip['Deltadp']/ip['md_at_2GeV']\
+                                              + ip['Deltas']/ip['ms_at_2GeV'])
 
     def value_pion_pole(self):
         """ Return the coefficient of the pion pole
 
         The pion pole is given, in terms of the spatial momentum q, by q^2 / (q^2 + mpi0^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
-            return ip.mN * ip.mtilde * ip.gA * (1/ip.mu_at_2GeV - 1/ip.md_at_2GeV) / 2
+            return self.mN * self.mtilde * ip['gA'] * (1/ip['mu_at_2GeV'] - 1/ip['md_at_2GeV']) / 2
         if self.nucleon == 'n':
-            return - ip.mN * ip.mtilde * ip.gA * (1/ip.mu_at_2GeV - 1/ip.md_at_2GeV) / 2
+            return - self.mN * self.mtilde * ip['gA'] * (1/ip['mu_at_2GeV'] - 1/ip['md_at_2GeV']) / 2
 
     def value_eta_pole(self):
         """ Return the coefficient of the eta pole
 
         The eta pole is given, in terms of the spatial momentum q, by q^2 / (q^2 + meta^2)
         """
-        ip = Num_input()
+        ip = self.input_dict
+        self.mtilde = 1/(1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] + 1/ip['ms_at_2GeV'])
+        self.mN = (ip['mproton'] + ip['mneutron'])/2
 
         if self.nucleon == 'p':
-            return ip.mN * ip.mtilde * (ip.Deltaup + ip.Deltadp - 2*ip.Deltas) * (1/ip.mu_at_2GeV + 1/ip.md_at_2GeV - 2/ip.ms_at_2GeV) / 6
+            return self.mN * self.mtilde * (ip['Deltaup'] + ip['Deltadp'] - 2*ip['Deltas'])\
+                * (1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] - 2/ip['ms_at_2GeV']) / 6
         if self.nucleon == 'n':
-            return ip.mN * ip.mtilde * (ip.Deltaun + ip.Deltadn - 2*ip.Deltas) * (1/ip.mu_at_2GeV + 1/ip.md_at_2GeV - 2/ip.ms_at_2GeV) / 6
+            return self.mN * self.mtilde * (ip['Deltaun'] + ip['Deltadn'] - 2*ip['Deltas'])\
+                * (1/ip['mu_at_2GeV'] + 1/ip['md_at_2GeV'] - 2/ip['ms_at_2GeV']) / 6